Move some common code to RawlsSimulation
