+++ src/perf/simulations/deleteWorkspaces.scala
@@ -1,17 +1,11 @@
 package default
 
 import scala.concurrent.duration._
-import java.io._
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
-import io.gatling.jdbc.Predef._
 
-class deleteWorkspaces extends Simulation {
-
-  val lines = scala.io.Source.fromFile("../user-files/config.txt").getLines
-  val accessToken = lines.next
-  val numUsers = lines.next.toInt
+class deleteWorkspaces extends RawlsSimulation {
 
   val workspaceListPath = "WORKSPACE_LIST.tsv" //list of workspaces to delete
