+++ src/perf/simulations/listWorkspaces.scala
@@ -1,19 +1,11 @@
 package default
 
 import scala.concurrent.duration._
-import java.io._
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
-import io.gatling.jdbc.Predef._
 
-class listWorkspaces extends Simulation {
-
-  //Helpers to set up the run
-
-  val lines = scala.io.Source.fromFile("../user-files/config.txt").getLines
-  val accessToken = lines.next
-  val numUsers = lines.next.toInt
+class listWorkspaces extends RawlsSimulation {
 
   //The run itself
